Implement `File` and some datatypes for MIME types
Currently, there is a separation between the image media type and other
ones. The reason for that was due to the requirement of an image MIME
type for the outer CIP-0025, while the one inside individual `File`s is
allowed to be any of the MIME types (a smarter datatype management can
probably address this).

Also, the implemented `Image` datatype might be too excessive. An
alternate solution could be utilizing the `elm/mimetype` package.